second

+++ prac1/mpi_blas_gemm.c
@@ -41,7 +41,7 @@ int main(int argc, char *argv[]) {
 	      scanf("%d",&num);
 		  
 		  for (i=1; i<np;i++){
-			  MPI_Recv(&num,1,MPI_INT,i, 0, MPI_COMM_WORLD,&st);
+			  MPI_Recv(&num,1,MPI_DOUBLE,i, 0, MPI_COMM_WORLD,&st);
 			  res = num + res;
 		  }
 		  printf("El resultado es: %d\n",res);
@@ -51,7 +51,7 @@ int main(int argc, char *argv[]) {
 	else
 	{
 	     
-		  MPI_Send(&num,1,MPI_INT,0,0,MPI_COMM_WORLD);
+		  MPI_Send(&num,1,MPI_DOUBLE,0,0,MPI_COMM_WORLD);
 		 
 	}
 	MPI_Finalize();
